requirements.txt
django==3.1.7
elastic-enterprise-search==7.12.0
elastic-transport==7.12.0
pyyaml==5.4.1
